refactor location module.
