+++ src/bloqade/ir/location/location.py
@@ -1,14 +1,22 @@
 from bloqade.builder.typing import ScalarType
-from beartype.typing import List, Tuple, Optional, Union, TYPE_CHECKING
+from bloqade.builder.start import ProgramStart
+from bloqade.ir.scalar import Scalar, Literal, cast
+from bloqade.ir.tree_print import Printer
+
+from pydantic.dataclasses import dataclass
+from beartype.typing import List, Tuple, Generator, Union, Optional
+from beartype import beartype
+from enum import Enum
+from numpy.typing import NDArray
+from bloqade.visualization import get_atom_arrangement_figure
+from bloqade.visualization import display_ir
+
 from beartype.vale import Is
 from typing import Annotated
-from beartype import beartype
 from plum import dispatch
+import plotext as pltxt
+import sys
 import numpy as np
-from bloqade.ir.scalar import cast
-
-if TYPE_CHECKING:
-    from bloqade.ir.location.list import ListOfLocations
 
 
 def check_position_array(array):
@@ -30,7 +38,156 @@ PositionArray = Annotated[np.ndarray, Is[check_position_array]]
 BoolArray = Annotated[np.ndarray, Is[check_bool_array]]
 
 
-class TransformTrait:
+class SiteFilling(int, Enum):
+    filled = 1
+    vacant = 0
+
+    def print_node(self) -> str:
+        return self.name
+
+    def children(self) -> List:
+        return []
+
+
+@dataclass(frozen=True)
+class LocationInfo:
+    position: Tuple[Scalar, Scalar]
+    filling: SiteFilling
+
+    @beartype
+    @staticmethod
+    def create(position: Tuple[ScalarType, ScalarType], filled: bool):
+        if filled:
+            filling = SiteFilling.filled
+        else:
+            filling = SiteFilling.vacant
+
+        position = tuple(cast(ele) for ele in position)
+
+        return LocationInfo(position, filling)
+
+    def print_node(self) -> str:
+        return f"Location: {self.filling.name}"
+
+    def children(self) -> List[Scalar]:
+        return {"x": self.position[0], "y": self.position[1]}
+
+
+@dataclass(init=False)
+class AtomArrangement(ProgramStart):
+    def __str__(self) -> str:
+        def is_literal(x):
+            return isinstance(x, Literal)
+
+        plot_results = all(
+            is_literal(info.position[0]) and is_literal(info.position[1])
+            for info in self.enumerate()
+        )
+
+        if plot_results:
+            xs = [float(info.position[0].value) for info in self.enumerate()]
+            ys = [float(info.position[1].value) for info in self.enumerate()]
+            filling = [bool(info.filling.value) for info in self.enumerate()]
+            xs_filled = [x for x, filled in zip(xs, filling) if filled]
+            ys_filled = [y for y, filled in zip(ys, filling) if filled]
+            xs_vacant = [x for x, filled in zip(xs, filling) if not filled]
+            ys_vacant = [y for y, filled in zip(ys, filling) if not filled]
+
+            pltxt.clear_figure()
+            pltxt.limit_size(False, False)
+            pltxt.plot_size(80, 24)
+            pltxt.canvas_color("default")
+            pltxt.axes_color("default")
+            pltxt.ticks_color("white")
+            pltxt.title("Atom Positions")
+            pltxt.xlabel("x (um)")
+            pltxt.ylabel("y (um)")
+
+            marker = "."
+            if sys.stdout.encoding.lower().startswith("utf"):
+                marker = "dot"
+
+            pltxt.scatter(xs_filled, ys_filled, color=(100, 55, 255), marker=marker)
+            if len(xs_vacant) > 0:
+                pltxt.scatter(
+                    xs_vacant, ys_vacant, color="white", label="vacant", marker=marker
+                )
+
+            return pltxt.build()
+        else:
+            ph = Printer()
+            ph.print(self)
+            return ph.get_value()
+
+    def _repr_pretty_(self, p, cycle):
+        Printer(p).print(self, cycle)
+
+    def print_node(self) -> str:
+        return "AtomArrangement"
+
+    def children(self) -> List[LocationInfo]:
+        return list(self.enumerate())
+
+    def enumerate(self) -> Generator[LocationInfo, None, None]:
+        """enumerate all locations in the register."""
+        raise NotImplementedError
+
+    def figure(self, fig_kwargs=None, **assignments):
+        """obtain a figure object from the atom arrangement."""
+        return get_atom_arrangement_figure(self, fig_kwargs, **assignments)
+
+    def show(self, **assignments) -> None:
+        display_ir(self, assignments)
+
+    def rydberg_interaction(self, **assignments) -> NDArray:
+        """calculate the Rydberg interaction matrix.
+
+        Args:
+            **assignments: the values to assign to the variables in the register.
+
+        Returns:
+            NDArray: the Rydberg interaction matrix in the lower triangular form.
+
+        """
+
+        from bloqade.constants import RB_C6
+
+        # calculate the Interaction matrix
+        V_ij = np.zeros((self.n_sites, self.n_sites))
+        for i, site_i in enumerate(self.enumerate()):
+            pos_i = np.array([float(ele(**assignments)) for ele in site_i.position])
+
+            for j, site_j in enumerate(self.enumerate()):
+                if j >= i:
+                    break  # enforce lower triangular form
+
+                pos_j = np.array([float(ele(**assignments)) for ele in site_j.position])
+                r_ij = np.linalg.norm(pos_i - pos_j)
+
+                V_ij[i, j] = RB_C6 / r_ij**6
+
+        return V_ij
+
+    @property
+    def n_atoms(self) -> int:
+        """number of atoms (filled sites) in the register."""
+        raise NotImplementedError
+
+    @property
+    def n_sites(self) -> int:
+        """number of sites in the register."""
+        raise NotImplementedError
+
+    @property
+    def n_vacant(self) -> int:
+        """number of vacant sites in the register."""
+        raise NotImplementedError
+
+    @property
+    def n_dims(self) -> int:
+        """number of dimensions in the register."""
+        raise NotImplementedError
+
     @beartype
     def scale(self, scale: ScalarType):
         """
@@ -67,8 +224,6 @@ class TransformTrait:
             shows your geometry in your web browser
 
         """
-        from .list import ListOfLocations
-        from .base import LocationInfo
 
         scale = cast(scale)
         location_list = []
@@ -85,9 +240,6 @@ class TransformTrait:
     def _add_position(
         self, position: Tuple[ScalarType, ScalarType], filling: Optional[bool] = None
     ):
-        from .list import ListOfLocations
-        from .base import LocationInfo
-
         if filling is None:
             filling = True
 
@@ -102,9 +254,6 @@ class TransformTrait:
         position: List[Tuple[ScalarType, ScalarType]],
         filling: Optional[List[bool]] = None,
     ):
-        from .list import ListOfLocations
-        from .base import LocationInfo
-
         location_list = list(self.enumerate())
 
         assert (
@@ -236,8 +385,6 @@ class TransformTrait:
             - `...apply_defect_count(defect_counts).show()`:
                 shows your geometry in your web browser
         """
-        from .list import ListOfLocations
-        from .base import LocationInfo, SiteFilling
 
         location_list = []
         for location_info in self.enumerate():
@@ -316,8 +463,6 @@ class TransformTrait:
             - `...apply_defect_count(defect_counts).show()`:
             shows your geometry in your web browser
         """
-        from .list import ListOfLocations
-        from .base import LocationInfo, SiteFilling
 
         p = min(1, max(0, defect_probability))
         location_list = []
@@ -340,12 +485,125 @@ class TransformTrait:
         return ListOfLocations(location_list=location_list)
 
     def remove_vacant_sites(self):
-        from .base import SiteFilling
-        from .list import ListOfLocations
-
         new_locations = []
         for location_info in self.enumerate():
             if location_info.filling is SiteFilling.filled:
                 new_locations.append(location_info)
 
         return ListOfLocations(new_locations)
+
+
+@dataclass(init=False)
+class ParallelRegisterInfo(ProgramStart):
+    """Parallel Register"""
+
+    __match_args__ = ("_register", "_cluster_spacing")
+
+    register_locations: List[List[Scalar]]
+    register_filling: List[int]
+    shift_vectors: List[List[Scalar]]
+
+    @beartype
+    def __init__(self, atom_arrangement: AtomArrangement, cluster_spacing: ScalarType):
+        self._register = atom_arrangement
+        self._cluster_spacing = cast(cluster_spacing)
+
+        if atom_arrangement.n_atoms > 0:
+            # calculate bounding box
+            # of this register
+            location_iter = atom_arrangement.enumerate()
+            (x, y) = next(location_iter).position
+            x_min = x
+            x_max = x
+            y_min = y
+            y_max = y
+
+            for location_info in location_iter:
+                (x, y) = location_info.position
+                x_min = x.min(x_min)
+                x_max = x.max(x_max)
+                y_min = y.min(y_min)
+                y_max = y.max(y_max)
+
+            shift_x = (x_max - x_min) + cluster_spacing
+            shift_y = (y_max - y_min) + cluster_spacing
+
+            register_locations = [
+                list(location_info.position)
+                for location_info in atom_arrangement.enumerate()
+            ]
+            register_filling = [
+                location_info.filling.value
+                for location_info in atom_arrangement.enumerate()
+            ]
+            shift_vectors = [[shift_x, cast(0)], [cast(0), shift_y]]
+        else:
+            raise ValueError("No locations to parallelize.")
+
+        self.register_locations = register_locations
+        self.register_filling = register_filling
+        self.shift_vectors = shift_vectors
+        super().__init__(self)
+
+
+@dataclass()
+class ParallelRegister:
+    atom_arrangement: AtomArrangement
+    cluster_spacing: Scalar
+
+    @property
+    def info(self) -> ParallelRegisterInfo:
+        return ParallelRegisterInfo(self.atom_arrangement, self.cluster_spacing)
+
+
+@dataclass(init=False)
+class ListOfLocations(AtomArrangement):
+    location_list: List[LocationInfo]
+
+    @beartype
+    def __init__(
+        self,
+        location_list: List[Union[LocationInfo, Tuple[ScalarType, ScalarType]]] = [],
+    ):
+        self.location_list = []
+        for ele in location_list:
+            if isinstance(ele, LocationInfo):
+                self.location_list.append(ele)
+            else:
+                self.location_list.append(LocationInfo.create(ele, True))
+
+        if self.location_list:
+            self.__n_atoms = sum(
+                1 for loc in self.location_list if loc.filling == SiteFilling.filled
+            )
+            self.__n_sites = len(self.location_list)
+            self.__n_vacant = self.__n_sites - self.__n_atoms
+            self.__n_dims = len(self.location_list[0].position)
+        else:
+            self.__n_sites = 0
+            self.__n_atoms = 0
+            self.__n_dims = None
+
+        super().__init__()
+
+    @property
+    def n_atoms(self):
+        return self.__n_atoms
+
+    @property
+    def n_sites(self):
+        return self.__n_sites
+
+    @property
+    def n_vacant(self):
+        return self.__n_vacant
+
+    @property
+    def n_dims(self):
+        return self.__n_dims
+
+    def enumerate(self):
+        return iter(self.location_list)
+
+    def __iter__(self):
+        return iter(self.location_list)
